Move components module to core
