+++ src/core/components/choice/mod.rs
@@ -1,5 +1,6 @@
 #[cfg(test)]
 mod tests;
+
 use std::collections::HashMap;
 
 use display::DisplayColor;
