+++ src/core/components/choice/tests.rs
@@ -3,7 +3,7 @@ use rstest::rstest;
 use view::assert_rendered_output;
 
 use super::*;
-use crate::components::testutil::handle_event_test;
+use crate::core::components::testutil::handle_event_test;
 
 #[derive(Clone, Debug, PartialEq)]
 pub enum TestAction {
